requirements.txt
streamlit==1.11.0
numpy==1.22.3
scikit-image==0.19.2
tensorflow-gpu==2.6.0
joblib==1.1.0
